+++ distributions/ssambanova/compose.yaml
@@ -1,13 +1,13 @@
 services:
   llamastack:
-    image: llamastack/distribution-ssambanova
+    image: llamastack/distribution-sambanova
     network_mode: "host"
     volumes:
       - ~/.llama:/root/.llama
-      - ./run.yaml:/root/llamastack-run-ssambanova.yaml
+      - ./run.yaml:/root/llamastack-run-sambanova.yaml
     ports:
       - "5000:5000"
-    entrypoint: bash -c "python -m llama_stack.distribution.server.server --yaml_config /root/llamastack-run-ssambanova.yaml"
+    entrypoint: bash -c "python -m llama_stack.distribution.server.server --yaml_config /root/llamastack-run-sambanova.yaml"
     deploy:
       restart_policy:
         condition: on-failure
