+++ llama_stack/providers/registry/inference.py
@@ -153,12 +153,12 @@ def available_providers() -> List[ProviderSpec]:
         remote_provider_spec(
             api=Api.inference,
             adapter=AdapterSpec(
-                adapter_type="ssambanova",
+                adapter_type="sambanova",
                 pip_packages=[
                     "openai",
                 ],
-                module="llama_stack.providers.remote.inference.ssambanova",
-                config_class="llama_stack.providers.remote.inference.ssambanova.SsambanovaImplConfig",
+                module="llama_stack.providers.remote.inference.sambanova",
+                config_class="llama_stack.providers.remote.inference.sambanova.SambanovaImplConfig",
             ),
         ),
     ]
