+++ llama_stack/templates/sambanova/__init__.py
@@ -4,4 +4,4 @@
 # This source code is licensed under the terms described in the LICENSE file in
 # the root directory of this source tree.
 
-from .ssambanova import get_distribution_template  # noqa: F401
+from .sambanova import get_distribution_template  # noqa: F401
